+++ README.md
@@ -100,7 +100,8 @@ import os
 BASE_DIR = './'                                                                  
 # Used internally by Django, can be anything of your choice                      
 SECRET_KEY = '<random string>'                                                   
-# API hostname, e.g. https://api.openbankproject.com                             
+# API hostname, e.g. https://api.openbankproject.com 
+# API OBP URL                            
 API_HOST = 'http://127.0.0.1:8080'                                                           
 # Consumer key + secret to authenticate the _app_ against the API                
 OAUTH_CONSUMER_KEY = '<key>'
